requirements.txt
fastapi
uvicorn
pydantic
python-multipart

pillow
pymupdf

paddlepaddle==2.6.2
paddleocr==2.7.0.3